# contracts/http/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Vinext AI Starter API
  version: 0.0.0
...
paths:
...
  /api/v1/tasks:
    post:
      operationId: createTask
      description: Create a Task and queue its processing. Requires a verified email and an Idempotency-Key header; repeating the key with the same payload replays the first response, and a different payload returns 409 with code idempotency_key_reused.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '202':
          description: The request has been accepted for processing, but processing has not yet completed.
          headers:
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  type:
                    type: string
                  title:
                    type: string
                  status:
                    type: integer
                    format: int32
                  detail:
                    type: string
                  code:
                    type: string
                required:
                  - type
                  - title
                  - status
                  - code
        '422':
          description: Client error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblem'
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
    get:
      operationId: listTasks
      description: Cursor-paginated list of the authenticated user's Tasks, newest first.
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          explode: false
        - name: per_page
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPage'
        '401':
          description: Access is unauthorized.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Access is forbidden.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      tags:
        - Tasks
...
components:
  schemas:
    CreateTaskRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
          maxLength: 10000
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
      description: RFC 9457 problem details.
...
    Task:
      type: object
      required:
        - id
        - input
        - output
        - state
        - version
        - error_code
        - correlation_id
        - started_at
        - finished_at
        - created_at
      properties:
        id:
          type: string
          description: Opaque identifier. Clients must not infer meaning from its value.
        input:
          type: string
        output:
          anyOf:
            - $ref: '#/components/schemas/TaskOutput'
            - type: 'null'
        state:
          $ref: '#/components/schemas/TaskState'
        version:
          type: integer
          format: int32
        error_code:
          anyOf:
            - type: string
            - type: 'null'
        correlation_id:
          type: string
        started_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        finished_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
        created_at:
          type: string
          format: date-time
      description: Reference asynchronous resource. The persisted state is the source of truth.
    TaskOutput:
      type: object
      required:
        - word_count
        - reversed
      properties:
        word_count:
          type: integer
          format: int32
        reversed:
          type: string
    TaskPage:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          type: object
          properties:
            next_cursor:
              anyOf:
                - type: string
                - type: 'null'
            prev_cursor:
              anyOf:
                - type: string
                - type: 'null'
          required:
            - next_cursor
            - prev_cursor
    TaskState:
      type: string
      enum:
        - queued
        - processing
        - completed
        - failed
...
    ValidationProblem:
      type: object
      required:
        - type
        - title
        - status
        - errors
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        errors:
          type: object
          unevaluatedProperties:
            type: array
            items:
              type: string
      description: RFC 9457 problem details with a stable map of field validation errors.