# packages/api-client/src/generated/models/taskPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Vinext AI Starter API
 * OpenAPI spec version: 0.0.0
 */
import type { Task } from "./task";
import type { TaskPageMeta } from "./taskPageMeta";

export interface TaskPage {
  data: Task[];
  meta: TaskPageMeta;
}